feat: migrate to PostgreSQL with major infrastructure modernization
This comprehensive update migrates the entire application from SQLite to
PostgreSQL while modernizing the dependency stack, API layer, and test
infrastructure.

## Database Migration to PostgreSQL

### Alembic Environment (env.py)
- Simplify database connection handling for PostgreSQL
- Remove SQLite-specific shared memory engine creation
- Import Base directly to avoid circular imports
- Update transaction handling for PostgreSQL compatibility
- Remove dialect_name overrides and SQLite PRAGMA statements

### All 22 Migration Files Updated
- Remove all SQLite-specific PRAGMA foreign_keys=OFF statements
- Replace sqlite_master queries with pg_indexes for introspection
- Convert CHECK constraints to native PostgreSQL ENUM types for ContentType
- Replace COLLATE NOCASE with functional LOWER() indexes
- Use batch_alter_table operations instead of table recreation
- Add proper identifier quoting for PostgreSQL reserved words
- Add constraint existence checking before drop/create operations
- Improve foreign key constraint handling with CASCADE options
- Use SQLAlchemy table/column constructs for data migrations

### New Migration
- ebb4481bb4c7: Change is_downloaded from Integer to Boolean (PostgreSQL-specific)
  - Use USING clause for proper type conversion
  - Handle server default changes during type migration

## Dependency Overhaul

### Added Dependencies
- **PostgreSQL drivers**: psycopg2-binary (>=2.9.10), asyncpg (>=0.30.0), psycopg[binary,pool] (>=3.2.10)
- **httpx-retries** (^0.4.3): Automatic retry handling for HTTP requests
- **ffmpeg-python** (^0.2.0): Replacing pyffmpeg for video processing
- **factory-boy** (^3.3.3): Test data generation (dev dependency)

### Removed Dependencies
- **requests**: Replaced with httpx for better async support
- **pyffmpeg** (2.4.2.18.1): Replaced with ffmpeg-python
- **tqdm**: Progress tracking consolidated into Rich
- **aiohttp**: Functionality migrated to httpx

### Updated Dependencies
- **httpx**: Now with http2 extras (^0.28.1)
- **gql**: Now with httpx and websockets extras (>=3.5.0)

### Pre-commit Hook Updates
- pre-commit-hooks: v5.0.0 → v6.0.0
- black: 24.10.0 → 25.9.0
- pyupgrade: v3.19.1 → v3.21.0
- flake8: 7.1.1 → 7.3.0
- isort: 5.13.2 → 7.0.0
- bandit: 1.8.2 → 1.8.6

## API Layer Refactor

### api/fansly.py
- Replace requests library with httpx for async HTTP support
- Add httpx_retries with configurable RetryTransport
- Integrate RateLimiter class for API rate limiting
- Update all HTTP method calls to httpx syntax
- Improve error handling with httpx exceptions
- Add proper timeout handling and retry logic

### api/rate_limiter.py (NEW)
- Implement token bucket algorithm for API rate limiting
- Async-first design with asyncio primitives
- Configurable rate limits and burst capacity

## Application Code Updates

### Database Layer (metadata/)
- Update all database operations for PostgreSQL compatibility
- Fix timezone handling for PostgreSQL TIMESTAMP WITH TIME ZONE
- Update query syntax for PostgreSQL-specific features
- Improve transaction handling and connection pooling
- Add asyncpg-specific optimizations

### Configuration (config/)
- Add PostgreSQL connection string support
- Update database URL validation for postgres:// scheme
- Add connection pool configuration options
- Update logging for database connection changes

### Download Layer (download/)
- Update progress tracking (tqdm → Rich)
- Improve async download handling with httpx
- Better error recovery with retry logic
- Update m3u8 download handling

### File I/O (fileio/, helpers/)
- Delete helpers/ffmpeg.py (pyffmpeg-based)
- Add helpers/rich_progress.py for unified progress tracking
- Update dedupe.py for PostgreSQL hash storage
- Improve video hashing with ffmpeg-python

### Stash Integration (stash/)
- Update all GraphQL queries for PostgreSQL data types
- Improve async processing with httpx-based client
- Update batch processing for better performance
- Fix metadata synchronization edge cases

## Test Infrastructure Overhaul

### Test Fixtures Reorganization
- Consolidate fixtures from scattered conftest.py into centralized modules:
  - tests/fixtures/api_fixtures.py
  - tests/fixtures/cleanup_fixtures.py
  - tests/fixtures/database_fixtures.py (moved from tests/metadata/conftest.py)
  - tests/fixtures/metadata_factories.py (NEW - factory-boy factories)
  - tests/fixtures/stash_api_fixtures.py (moved from tests/stash/conftest.py)
  - tests/fixtures/stash_integration_fixtures.py
  - tests/fixtures/stash_mixin_fixtures.py
  - tests/fixtures/stash_processing_fixtures.py

### Factory-Boy Integration
- Replace Mock objects with factory-boy factories for SQLAlchemy models
- Add factories for Account, Media, Post, Message, MediaLocation, etc.
- Enable more realistic test data generation
- Improve test maintainability and readability

### Test Configuration Updates (pyproject.toml)
- Add parallel test execution with pytest-xdist (-n=8 workers)
- Update coverage configuration for PostgreSQL
- Add filter warnings for SQLAlchemy, coverage, pluggy
- Enable multiprocessing concurrency
- Improve coverage reporting with htmlcov output

### Test File Updates (83 test files modified)
- Update all database tests for PostgreSQL syntax
- Replace mock-heavy tests with fixture-based approaches
- Add proper async database session handling
- Update assertions for PostgreSQL-specific behavior
- Remove SQLite-specific test cases
- Add PostgreSQL-specific edge case tests

## Project Metadata

### pyproject.toml Structure
- Migrate from tool.poetry format to modern [project] format
- Move build-system to top of file per PEP 518
- Update authors list with proper name/email format
- Add credits section for contributors
- Update Poetry core requirement to >=2.0.0,<3.0.0

## Documentation

### New Documentation Files
- docs/SxS-PostgreSQL-Implementation-Plan.md: Side-by-side migration strategy
- docs/hardCut-PostgreSQL-Implementation-Plan.md: Hard cutover migration plan

### Updated .gitignore
- Add Snyk Security Extension exclusions
- Add PostgreSQL-specific temporary files

## Migration Notes

- PostgreSQL enforces foreign keys by default (no PRAGMA needed)
- Application handles data integrity at business logic level to accommodate
  API data that may arrive in non-standard order
- Some enum downgrade operations use pass (manual handling required)
- All migrations tested for both upgrade and downgrade paths
- Compatible with PostgreSQL 12+ (tested with 14+)

## Breaking Changes

- **Database**: SQLite databases must be migrated to PostgreSQL using
  provided migration scripts (see docs/hardCut-PostgreSQL-Implementation-Plan.md)
- **Configuration**: Database connection strings must use postgres:// scheme
- **Dependencies**: Requires PostgreSQL server installation
- **Python**: Minimum version remains 3.12

## Performance Improvements

- Better connection pooling with psycopg3
- Async database operations with asyncpg
- HTTP/2 support via httpx
- Parallel test execution (8 workers)
- Improved retry logic reduces transient failures

## Testing

All 156 modified files have been updated with:
- PostgreSQL-compatible assertions
- Factory-based test data generation
- Proper async context management
- Comprehensive edge case coverage

+++ metadata/attachable.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import BigInteger, Column, Integer, String
 from sqlalchemy.orm import Mapped, declared_attr, mapped_column
 
 from .base import Base
@@ -26,7 +26,7 @@ class Attachable(Base):
     def __tablename__(self) -> str:
         return self.__name__.lower()
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
     type: Mapped[str] = mapped_column(String(50))
 
     __mapper_args__ = {
